// irmin-console/src/types/Wordpress.ts
export interface ArticlesSection {
  acf_fc_layout: 'articles';
  title: string;
  subtitle: string;
  description: string;
}